Update My_Project.py
Could not get the raw_input working, but I did change all the inputs to strings.
Added a function to check off to-do list items, it is currently imbeded in the while loop but I may also put the option at the end when the list is finalized.
I also need to find a way to delete the corresponding time from the time list when an item is checked off.  I think creating a dictionary may help with this.